Interface refactor
