+++ src/main/java/com/myfinancemap/app/service/interfaces/TransactionService.java
@@ -1,7 +1,7 @@
-package com.myfinancemap.app.service;
+package com.myfinancemap.app.service.interfaces;
 
-import com.myfinancemap.app.dto.TotalCostResponse;
-import com.myfinancemap.app.dto.TransactionDto;
+import com.myfinancemap.app.dto.transaction.TotalCostResponse;
+import com.myfinancemap.app.dto.transaction.TransactionDto;
 
 import java.time.LocalDate;
 import java.util.List;
